+++ velox/connectors/clp/search_lib/archive/ClpArchiveVectorLoader.cpp
@@ -21,6 +21,7 @@
 #include "clp_s/ColumnReader.hpp"
 #include "clp_s/SchemaTree.hpp"
 #include "velox/connectors/clp/search_lib/BaseClpCursor.h"
+#include "velox/connectors/clp/search_lib/ClpTimestampsUtils.h"
 #include "velox/connectors/clp/search_lib/archive/ClpArchiveVectorLoader.h"
 #include "velox/type/Timestamp.h"
 #include "velox/vector/ComplexVector.h"
